update for run

+++ minee/model/mine.py
@@ -123,7 +123,7 @@ class Mine():
             log.write("video_rate={0}\n".format(self.video_rate))
             log.write("infinite_sample={0}\n".format(self.infinite_sample))
             log.close()
-            if self.googleDrive:
+            if self.googleDrive and self.prefixID:
                 self.googleDrive.uploadFile(log_file, "{}_train.log".format(self.model_name), self.prefixID)
             
 
@@ -234,7 +234,7 @@ class Mine():
                         torch.save(self.state_dict(),f)
                         if self.verbose:
                             print('results saved: '+str(snapshot_i))
-                    if self.googleDrive:
+                    if self.googleDrive and self.prefixID:
                         self.googleDrive.uploadFile(fname_i, "cache_iter={}.pt".format(i+1), parentID=self.prefixID)
                         os.remove(fname_i)
                 snapshot_i += 1
@@ -304,7 +304,11 @@ class Mine():
         fpath = os.path.join(self.prefix, "archive")
         if not os.path.exists(fpath):
             os.mkdir(fpath)
-            self.archiveID = self.googleDrive.createFolder("archive", self.prefixID)
+        if self.googleDrive and self.prefixID:
+            self.archiveID = self.googleDrive.searchFolder("archive", parentID=self.prefixID)
+            if not self.archiveID:
+                self.archiveID = self.googleDrive.createFolder("archive", self.prefixID)
+            # self.archiveID = self.googleDrive.createFolder("archive", self.prefixID)
         if self.video_rate>0:
             self.save_video(array_end)
         fname__ = "[{}-{}).pt".format(self.array_start, array_end)
@@ -323,7 +327,7 @@ class Mine():
             if self.video_rate>0:
                 for j in range(self.rep):
                     self.ixy_list[j] = np.zeros(self.ixy_list_shape)
-        if self.googleDrive:
+        if self.googleDrive and self.archiveID:
             self.googleDrive.uploadFile(fname,  fname__, self.archiveID)
             os.remove(fname)
 
@@ -460,7 +464,7 @@ class Mine():
         figName = os.path.join(self.prefix, "{}_{}.png".format(self.model_name, suffix))
         fig.savefig(figName, bbox_inches='tight')
         plt.close()
-        if self.googleDrive:
+        if self.googleDrive and self.prefixID:
             self.googleDrive.uploadFile(figName,   "{}_{}.png".format(self.model_name, suffix), self.prefixID)
             os.remove(figName)
